adf4350_sink: Rework the block structure.
Because of some modifications in the libm2k communication segment, ADF4350 block was updated, in order to match with the new structure.
