+++ lib/drivers/adf4350.h
@@ -44,7 +44,7 @@
 /***************************** Include Files **********************************/
 /******************************************************************************/
 #include <stdint.h>
-#include "communication/spi/spi.h"
+#include "spi.h"
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
@@ -130,70 +130,70 @@
 /************************ Types Definitions ***********************************/
 /******************************************************************************/
 struct adf4350_platform_data {
-    uint32_t	clkin;
-    uint32_t	channel_spacing;
-    uint64_t	power_up_frequency;
+	uint32_t	clkin;
+	uint32_t	channel_spacing;
+	uint64_t	power_up_frequency;
 
-    uint16_t	ref_div_factor; /* 10-bit R counter */
-    uint8_t	    ref_doubler_en;
-    uint8_t	    ref_div2_en;
+	uint16_t	ref_div_factor; /* 10-bit R counter */
+	uint8_t	    ref_doubler_en;
+	uint8_t	    ref_div2_en;
 
-    uint32_t    r2_user_settings;
-    uint32_t    r3_user_settings;
-    uint32_t    r4_user_settings;
-    int32_t	    gpio_lock_detect;
+	uint32_t    r2_user_settings;
+	uint32_t    r3_user_settings;
+	uint32_t    r4_user_settings;
+	int32_t	    gpio_lock_detect;
 };
 
 typedef struct {
-    /* SPI */
-    spi_init_param	spi_init;
+	/* SPI */
+	spi_init_param	spi_init;
 
-    /* Device settings */
-    uint32_t	clkin;
-    uint32_t	channel_spacing;
-    uint32_t	power_up_frequency;
-    uint32_t	reference_div_factor;
-    uint8_t		reference_doubler_enable;
-    uint8_t		reference_div2_enable;
+	/* Device settings */
+	uint32_t	clkin;
+	uint32_t	channel_spacing;
+	uint32_t	power_up_frequency;
+	uint32_t	reference_div_factor;
+	uint8_t		reference_doubler_enable;
+	uint8_t		reference_div2_enable;
 
-    /* r2_user_settings */
-    uint8_t		phase_detector_polarity_positive_enable;
-    uint8_t		lock_detect_precision_6ns_enable;
-    uint8_t		lock_detect_function_integer_n_enable;
-    uint32_t	charge_pump_current;
-    uint32_t	muxout_select;
-    uint8_t		low_spur_mode_enable;
+	/* r2_user_settings */
+	uint8_t		phase_detector_polarity_positive_enable;
+	uint8_t		lock_detect_precision_6ns_enable;
+	uint8_t		lock_detect_function_integer_n_enable;
+	uint32_t	charge_pump_current;
+	uint32_t	muxout_select;
+	uint8_t		low_spur_mode_enable;
 
-    /* r3_user_settings */
-    uint8_t		cycle_slip_reduction_enable;
-    uint8_t		charge_cancellation_enable;
-    uint8_t		anti_backlash_3ns_enable;
-    uint8_t		band_select_clock_mode_high_enable;
-    uint32_t	clk_divider_12bit;
-    uint32_t	clk_divider_mode;
+	/* r3_user_settings */
+	uint8_t		cycle_slip_reduction_enable;
+	uint8_t		charge_cancellation_enable;
+	uint8_t		anti_backlash_3ns_enable;
+	uint8_t		band_select_clock_mode_high_enable;
+	uint32_t	clk_divider_12bit;
+	uint32_t	clk_divider_mode;
 
-    /* r4_user_settings */
-    uint8_t		aux_output_enable;
-    uint8_t		aux_output_fundamental_enable;
-    uint8_t		mute_till_lock_enable;
-    uint32_t	output_power;
-    uint32_t	aux_output_power;
+	/* r4_user_settings */
+	uint8_t		aux_output_enable;
+	uint8_t		aux_output_fundamental_enable;
+	uint8_t		mute_till_lock_enable;
+	uint32_t	output_power;
+	uint32_t	aux_output_power;
 } adf4350_init_param;
 
 typedef struct {
-    spi_desc_t	*spi_desc;
-    struct adf4350_platform_data *pdata;
-    uint32_t	clkin;
-    uint32_t	chspc;	/* Channel Spacing */
-    uint32_t	fpfd;	/* Phase Frequency Detector */
-    uint32_t	min_out_freq;
-    uint32_t	r0_fract;
-    uint32_t	r0_int;
-    uint32_t	r1_mod;
-    uint32_t	r4_rf_div_sel;
-    uint32_t	regs[6];
-    uint32_t	regs_hw[6];
-    uint32_t 	val;
+	spi_desc	*spi_desc;
+	struct adf4350_platform_data *pdata;
+	uint32_t	clkin;
+	uint32_t	chspc;	/* Channel Spacing */
+	uint32_t	fpfd;	/* Phase Frequency Detector */
+	uint32_t	min_out_freq;
+	uint32_t	r0_fract;
+	uint32_t	r0_int;
+	uint32_t	r1_mod;
+	uint32_t	r4_rf_div_sel;
+	uint32_t	regs[6];
+	uint32_t	regs_hw[6];
+	uint32_t 	val;
 } adf4350_dev;
 
 /******************************************************************************/
@@ -201,21 +201,21 @@ typedef struct {
 /******************************************************************************/
 /*! Initializes the ADF4350. */
 int32_t adf4350_setup(adf4350_dev **device,
-                      adf4350_init_param init_param);
+		      adf4350_init_param init_param);
 /*! Writes 4 bytes of data to ADF4350. */
 int32_t adf4350_write(adf4350_dev *dev,
-                      uint32_t data);
+		      uint32_t data);
 /*! Stores PLL 0 frequency in Hz. */
 int64_t adf4350_out_altvoltage0_frequency(adf4350_dev *dev,
-                                          int64_t Hz);
+					  int64_t Hz);
 /*! Stores PLL 0 frequency resolution/channel spacing in Hz. */
 int32_t adf4350_out_altvoltage0_frequency_resolution(adf4350_dev *dev,
-                                                     int32_t Hz);
+						     int32_t Hz);
 /*! Sets PLL 0 REFin frequency in Hz. */
 int64_t adf4350_out_altvoltage0_refin_frequency(adf4350_dev *dev,
-                                                int64_t Hz);
+						int64_t Hz);
 /*! Powers down the PLL.  */
 int32_t adf4350_out_altvoltage0_powerdown(adf4350_dev *dev,
-                                          int32_t pwd);
+					  int32_t pwd);
 
 #endif // __ADF4350_H__
