--- lib/drivers/delay.cpp
@@ -1,7 +1,0 @@
-#include "delay.h"
-#include <unistd.h>
-
-
-void mdelay(uint32_t msecs) {
-    usleep(msecs * 1000);
-}
